[UPDATE] 委譲 (カット・アンド・ペースト)
